+++ packages/governance/src/plugin/ai-payload-limits.ts
@@ -1,0 +1,9 @@
+export const AI_PAYLOAD_LIMITS = {
+  rootCauseDependencies: 120,
+  prImpactDependencies: 120,
+  driftSignals: 12,
+  driftDeltas: 12,
+  driftViolations: 20,
+  scorecardViolations: 20,
+  scorecardDeltas: 12,
+} as const;
